Can now get SD cards size properly on format, reduce offsets from 64 to 32 bit
If theres a > 4gb savestate, thats the problem, not the address bits in the offset.

+++ src/ads7846.h
@@ -7,7 +7,7 @@
 extern bool ads7846PenIrqEnabled;
 
 void ads7846Reset(void);
-uint64_t ads7846StateSize(void);
+uint32_t ads7846StateSize(void);
 void ads7846SaveState(uint8_t* data);
 void ads7846LoadState(uint8_t* data);
